Suggested update
I'd suggest merging the methods setHealth, addHealth, and subHealth into a single method, especially since the setHealth method is pretty much redundant at the moment. (A negative value can be passed to the new updateHealth method if health must be taken off.)

Also, there shouldn't really be any need for the setStrength method, since it should only need to be set once. If you're looking to find out how much more a player can carry in terms of weight, then it would be better to get the total weight of the inventory, and compare this to his strength:

public void addObject(MovableObject object)
{
    int canCarry = getStrength() - getInventoryWeight();

    if(canCarry = 0) {
        // return can't carry anything else
    }

    if(canCarry - object.getWeight() < 0) {
        // return can't carry object, not enough strength left
    }

    // add item to inventory
    // return item added to inventory
}